src/test/java/com/godaddy/sonar/ruby/metricfu/MetricfuComplexitySensorTest.java
package com.godaddy.sonar.ruby.metricfu;

import static org.easymock.EasyMock.expect;
import static org.easymock.EasyMock.isA;
import static org.junit.Assert.assertNotNull;

import java.io.File;
import java.io.IOException;
import java.util.ArrayList;
import java.util.List;

import org.apache.commons.configuration.Configuration;
import org.easymock.EasyMock;
import org.easymock.IMocksControl;
import org.junit.Before;
import org.junit.Test;
import org.sonar.api.batch.SensorContext;
import org.sonar.api.batch.fs.FileSystem;
import org.sonar.api.config.Settings;
import org.sonar.api.resources.Project;
import org.sonar.api.scan.filesystem.PathResolver;

import com.godaddy.sonar.ruby.core.LanguageRuby;


public class MetricfuComplexitySensorTest
{
	private IMocksControl mocksControl;

	private PathResolver pathResolver;
	private SensorContext sensorContext;
	private MetricfuYamlParser metricfuYamlParser;
	private MetricfuComplexitySensor metricfuComplexitySensor;
	private Configuration config;
	private Project project;

	private Settings settings;
	private FileSystem fs;

	@Before
	public void setUp() throws Exception
	{
		mocksControl = EasyMock.createControl();
		pathResolver = mocksControl.createMock(PathResolver.class);
		fs = mocksControl.createMock(FileSystem.class);
		metricfuYamlParser = mocksControl.createMock(MetricfuYamlParser.class);
		settings = new Settings();

		metricfuComplexitySensor = new MetricfuComplexitySensor(settings, fs, metricfuYamlParser);
		config = mocksControl.createMock(Configuration.class);
		expect(config.getString("sonar.language", "java")).andStubReturn("ruby");

		project = new Project("test project");
		project.setLanguage(LanguageRuby.INSTANCE);
//		project.setConfiguration(config);

	}

	@Test
	public void testConstructor()
	{
		assertNotNull(metricfuComplexitySensor);
	}

	@Test
	public void testAnalyse() throws IOException
	{
		List<File> sourceFiles= new ArrayList<File>();
		List<File> sourceDirs = new ArrayList<File>();

		sourceDirs.add(new File("lib"));
		sourceFiles.add(new File("lib/some_path/foo_bar.rb"));

		sensorContext = mocksControl.createMock(SensorContext.class);

		expect(fs.baseDir()).andReturn(new File("bar"));
        expect(pathResolver.relativeFile(isA(File.class),isA(String.class))).andReturn(new File("foo"));
		mocksControl.replay();

//		metricfuComplexitySensor.analyse(project, sensorContext);
//		mocksControl.verify();
	}
}
